backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.23
asyncpg==0.28.0
aiosqlite==0.19.0
pydantic==2.6.1
pydantic-settings==2.2.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
authlib==1.3.0
httpx==0.27.0
openpyxl==3.1.2

